refactoring code

+++ chapter_1/fahr_to_celsius.c
@@ -3,14 +3,14 @@
 #include <stdio.h>
 
 int main(int argc, char *argv[]){
-  int celsius;
+  float celsius;
   int fahr = 0;
   int limit = 300;
   int step = 20;
 
   for(fahr = 0; fahr <= limit ; fahr=fahr+step){
     celsius = (5.0/9.0) * (fahr - 32);
-    printf("%i\t%i\n", fahr, celsius);
+    printf("%3d\t%6.1f\n", fahr, celsius);
   }
   return 0;
 }
